Generalise test job request unauthorised user
